requirements.txt
aiogram==2.23.1
python-dotenv==0.21.0
python-telegram-bot==13.15
openpyxl==3.0.10

#python-decouple==3.6
#telegram==0.0.1
#tornado==6.2
#lxml==4.9.1
#requests~=2.28.1